First evolution!
Added entity energy, governing how long they can live, when they can
reproduce, etc, with the loss of energy based on the stats, allowing
for the first basic evolution! It is still early days, and is rather
predictable (after watching it only a few times, I know the standard
trends already), but it is something. There are a number of bugs
remaining though, which I won't fix now, because it is 12 am and I
am feeling less human by the second, but the currently 'known' 'bugs'
(the quotes will make sense after reading them) are as follows:

- "Error: main.lua:69: loop in gettable" crash sometimes after longer
  periods of running, however it might be fixed due to commenting
  the 'teleport to food if close' lines, this needs to be tested
  though
- Sometimes children don't seem to appear from their parent, this
  might be sleep deprivation though
- Not really a bug, but further tweaking the energy loss equation
  may produce more varied and interesting outcomes, or at least
  not just 'small yellow, then eliminate competition, then get bigger'